Some more fiddling

+++ init.sh
@@ -3,13 +3,48 @@
 # Use of this source code is governed by the MIT license, a copy of which can
 # be found in the LICENSE file.
 
-ln -s ../playground/data data
-ln -s ../playground/javascript javascript
+chmod +x samp03svr samp-npc announce
 
-ln -s ../echo-plugin/src/out/echo-plugin.so plugins/echo-plugin.so
-ln -s ../json-plugin/src/out/json-plugin.so plugins/json-plugin.so
+if [ ! -f server.cfg-private ]; then
+  echo "You must have a server.cfg-private file to initialize the LVP server."
+  exit
+fi
+
+if [ ! -f data ]; then
+  ln -s ../playground/data data
+fi
+
+if [ ! -f data ]; then
+  ln -s ../playground/javascript javascript
+fi
+
+if [ ! -f plugins/echo-plugin.so ]; then
+  ln ../echo-plugin/src/out/echo-plugin.so plugins/echo-plugin.so
+fi
+
+if [ ! -f plugins/json-plugin.so ]; then
+  ln ../json-plugin/src/out/json-plugin.so plugins/json-plugin.so
+fi
+
 # TODO: mysql-plugin.so
-ln -s ../playgroundjs-plugin/src/out/playgroundjs-plugin.so plugins/playgroundjs-plugin.so
-ln -s ../playgroundjs-plugin/src/out/libv8.so plugins/libv8.so
-# TODO: streamer.so
-ln -s ../zone-manager-plugin/src/out/zone-manager-plugin.so plugins/zone-manager-plugin.so
+
+if [ ! -f plugins/playgroundjs-plugin.so ]; then
+  ln ../playgroundjs-plugin/src/out/playgroundjs-plugin.so plugins/playgroundjs-plugin.so
+fi
+
+if [ ! -f plugins/libv8.so ]; then
+  ln ../playgroundjs-plugin/src/out/libv8.so libv8.so
+fi
+
+if [ ! -f plugins/streamer.so ]; then
+  ln ../samp-streamer-plugin/bin/linux/Release/streamer.so plugins/streamer.so
+fi
+
+if [ ! -f plugins/zone-manager-plugin.so ]; then
+  ln ../zone-manager-plugin/src/out/zone-manager-plugin.so plugins/zone-manager-plugin.so
+fi
+
+# Create server.cfg by combining server.cfg-base and server.cfg-private
+source server.cfg-private
+
+eval "echo \"$(< server.cfg-base)\"" > server.cfg
